requirements.txt
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
